fmt + clippy
